Arreglado el metodo de simulacion para que ahora pare cuando se produzca la colision

+++ PracticaCinematica/Constantes.h
@@ -1,6 +1,6 @@
 #pragma once
 namespace constantes
 {
-	const float PI = 3.14159265;
-	const float GRAVEDAD = 9.80665;
+	const double PI = 3.14159265;
+	const double GRAVEDAD = 9.80665;
 }
